+++ 3rdparty/qwt/src/qwt_arrow_button.cpp
@@ -8,12 +8,12 @@
  *****************************************************************************/
 
 #include "qwt_arrow_button.h"
+#include "qwt.h"
 
 #include <qpainter.h>
 #include <qstyle.h>
 #include <qstyleoption.h>
 #include <qevent.h>
-#include <qapplication.h>
 
 static const int MaxNum = 3;
 static const int Margin = 2;
@@ -29,7 +29,7 @@ public:
 static QStyleOptionButton styleOpt( const QwtArrowButton* btn )
 {
     QStyleOptionButton option;
-    option.init( btn );
+    option.initFrom( btn );
     option.features = QStyleOptionButton::None;
     if ( btn->isFlat() )
         option.features |= QStyleOptionButton::Flat;
@@ -198,7 +198,7 @@ void QwtArrowButton::drawButtonLabel( QPainter *painter )
     if ( hasFocus() )
     {
         QStyleOptionFocusRect option;
-        option.init( this );
+        option.initFrom( this );
         option.backgroundColor = palette().color( QPalette::Window );
 
         style()->drawPrimitive( QStyle::PE_FrameFocusRect,
@@ -260,7 +260,7 @@ void QwtArrowButton::drawArrow( QPainter *painter,
 QSize QwtArrowButton::sizeHint() const
 {
     const QSize hint = minimumSizeHint();
-    return hint.expandedTo( QApplication::globalStrut() );
+    return qwtExpandedToGlobalStrut( hint );
 }
 
 /*!
@@ -279,7 +279,7 @@ QSize QwtArrowButton::minimumSizeHint() const
         sz.transpose();
 
     QStyleOption styleOption;
-    styleOption.init( this );
+    styleOption.initFrom( this );
 
     sz = style()->sizeFromContents( QStyle::CT_PushButton,
         &styleOption, sz, this );