+++ 3rdparty/qwt/src/qwt_slider.cpp
@@ -12,13 +12,14 @@
 #include "qwt_scale_draw.h"
 #include "qwt_scale_map.h"
 #include "qwt_math.h"
+#include "qwt.h"
 
 #include <qevent.h>
 #include <qdrawutil.h>
 #include <qpainter.h>
 #include <qstyle.h>
 #include <qstyleoption.h>
-#include <qapplication.h>
+#include <qmargins.h>
 
 static QSize qwtHandleSize( const QSize &size,
     Qt::Orientation orientation, bool hasTrough )
@@ -677,7 +678,7 @@ void QwtSlider::paintEvent( QPaintEvent *event )
     painter.setClipRegion( event->region() );
 
     QStyleOption opt;
-    opt.init(this);
+    opt.initFrom(this);
     style()->drawPrimitive(QStyle::PE_Widget, &opt, &painter, this);
 
     if ( d_data->scalePosition != QwtSlider::NoScale )
@@ -911,7 +912,7 @@ bool QwtSlider::hasGroove() const
 QSize QwtSlider::sizeHint() const
 {
     const QSize hint = minimumSizeHint();
-    return hint.expandedTo( QApplication::globalStrut() );
+    return qwtExpandedToGlobalStrut( hint );
 }
 
 /*!
@@ -974,11 +975,10 @@ QSize QwtSlider::minimumSizeHint() const
     }
 
     // finally add margins
-    int left, right, top, bottom;
-    getContentsMargins( &left, &top, &right, &bottom );
+    const QMargins m = contentsMargins();
 
-    w += left + right;
-    h += top + bottom;
+    w += m.left() + m.right();
+    h += m.top() + m.bottom();
 
     d_data->sizeHintCache = QSize( w, h );
     return d_data->sizeHintCache;